+++ src/game/text.h
@@ -16,6 +16,8 @@ void Text_Flash(TEXTSTRING *textstring, bool enable, int16_t rate);
 void Text_Hide(TEXTSTRING *textstring, bool enable);
 void Text_AddBackground(
     TEXTSTRING *textstring, int16_t w, int16_t h, int16_t x, int16_t y);
+void Text_CentreVGradient(
+    TEXTSTRING *textstring, RGBA8888 centre, RGBA8888 edge);
 void Text_RemoveBackground(TEXTSTRING *textstring);
 void Text_AddOutline(TEXTSTRING *textstring, bool enable);
 void Text_RemoveOutline(TEXTSTRING *textstring);